# docs/swagger.yaml
basePath: /api/v1
definitions:
  entities.Allergy:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        example: Пенициллин
        type: string
      patient:
        items:
          $ref: '#/definitions/entities.Patient'
        type: array
      updated_at:
        type: string
    type: object
  entities.ContactInfo:
    properties:
      address:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      phone:
        type: string
      updated_at:
        type: string
    type: object
  entities.Doctor:
    properties:
      created_at:
        type: string
      emergency_calls:
        items:
          $ref: '#/definitions/entities.EmergencyCall'
        type: array
      id:
        type: integer
      login:
        example: doctor_ivanov
        type: string
      receptions:
        items:
          $ref: '#/definitions/entities.ReceptionHospital'
        type: array
      specialization:
        example: Терапевт
        type: string
      updated_at:
        type: string
    type: object
  entities.EmergencyCall:
    properties:
      address:
        example: Москва, ул. Ленина, д. 15
        type: string
      created_at:
        type: string
      doctor_id:
        example: 1
        type: integer
      id:
        type: integer
      phone:
        example: "+79991234567"
        type: string
      priority:
        example: true
        type: boolean
      receptions:
        items:
          $ref: '#/definitions/entities.ReceptionSMP'
        type: array
      status:
        allOf:
        - $ref: '#/definitions/entities.EmergencyStatus'
        example: in_progress
      updated_at:
        type: string
    type: object
  entities.EmergencyStatus:
    enum:
    - scheduled
    - accepted
    - on_place
    - completed
    - cancelled
    - no_show
    type: string
    x-enum-varnames:
    - EmergencyStatusScheduled
    - EmergencyStatusAccepted
    - EmergencyStatusOnPlace
    - EmergencyStatusCompleted
    - EmergencyStatusCancelled
    - EmergencyStatusNoShow
  entities.MedService:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        example: EKG
        type: string
      price:
        example: 100
        type: integer
      receptionSMP:
        items:
          $ref: '#/definitions/entities.ReceptionSMP'
        type: array
      updated_at:
        type: string
    type: object
  entities.Patient:
    properties:
      allergy:
        items:
          $ref: '#/definitions/entities.Allergy'
        type: array
      birth_date:
        example: "1980-05-15T00:00:00Z"
        type: string
      contact_info:
        $ref: '#/definitions/entities.ContactInfo'
      created_at:
        type: string
      emergency_receptions:
        items:
          $ref: '#/definitions/entities.ReceptionSMP'
        type: array
      full_name:
        example: Смирнов Алексей Петрович
        type: string
      id:
        type: integer
      is_male:
        example: true
        type: boolean
      personal_info:
        $ref: '#/definitions/entities.PersonalInfo'
      receptions:
        items:
          $ref: '#/definitions/entities.ReceptionHospital'
        type: array
      updated_at:
        type: string
    type: object
  entities.PersonalInfo:
    properties:
      created_at:
        type: string
      id:
        type: integer
      oms:
        example: "1234567890123456"
        type: string
      passport_series:
        example: 4510 123456
        type: string
      patient_id:
        example: 1
        type: integer
      snils:
        example: 123-456-789 00
        type: string
      updated_at:
        type: string
    type: object
  entities.ReceptionHospital:
    properties:
      address:
        example: Москва, ул. Ленина, д. 15
        type: string
      created_at:
        type: string
      date:
        example: "2023-10-15T14:30:00Z"
        type: string
      diagnosis:
        example: ОРВИ
        type: string
      doctor:
        $ref: '#/definitions/entities.Doctor'
      doctor_id:
        example: 1
        type: integer
      id:
        type: integer
      patient:
        $ref: '#/definitions/entities.Patient'
      patient_id:
        example: 1
        type: integer
      recommendations:
        example: Постельный режим
        type: string
      status:
        allOf:
        - $ref: '#/definitions/entities.ReceptionStatus'
        example: scheduled
      updated_at:
        type: string
    type: object
  entities.ReceptionSMP:
    properties:
      created_at:
        type: string
      diagnosis:
        example: ОРВИ
        type: string
      doctor_id:
        example: 1
        type: integer
      emergency_call_id:
        example: 1
        type: integer
      id:
        type: integer
      med_services:
        items:
          $ref: '#/definitions/entities.MedService'
        type: array
      patient_id:
        example: 1
        type: integer
      recommendations:
        example: Постельный режим
        type: string
      updated_at:
        type: string
    type: object
  entities.ReceptionStatus:
    enum:
    - scheduled
    - completed
    - cancelled
    - no_show
    type: string
    x-enum-comments:
      StatusCancelled: '"Отменен"'
      StatusCompleted: '"Завершен"'
      StatusNoShow: '"Не явился"'
      StatusScheduled: '"Запланирован"'
    x-enum-varnames:
    - StatusScheduled
    - StatusCompleted
    - StatusCancelled
    - StatusNoShow
  handlers.ResultError:
    properties:
      response:
        properties:
          code:
            description: '[RULE]: must be one of codes from table (Check DEV.PAGE)'
            example: 400
            type: integer
          message:
            example: Bad request
            type: string
        type: object
      status:
        description: error
        example: error
        type: string
    type: object
  handlers.ResultResponse:
    properties:
      response:
        properties:
          data:
            description: '[AVALIABLE]: object, array of objects, empty'
          message:
            example: Success operation
            type: string
          type:
            description: '[AVALIABLE]: object, array, empty'
            example: object
            type: string
        type: object
      status:
        description: ok
        example: ok
        type: string
    type: object
...
  models.CreatePatientRequest:
    description: Данные для создания нового пациента
    properties:
      birth_date:
        description: Дата рождения
        example: "1980-05-15"
        type: string
      full_name:
        description: ФИО пациента
        example: Смирнов Алексей Петрович
        type: string
      is_male:
        description: Пол (true - мужской)
        example: true
        type: boolean
    type: object
...
  models.ReceptionShortResponse:
    description: Содержит основные данные приема для отображения в списках
    properties:
      address:
        example: Москва, ул. Ленина, д. 15
        type: string
      date:
        description: Форматированная дата приема
        example: 15.10.2023 14:30
        type: string
      diagnosis:
        example: ОРВИ
        type: string
      id:
        type: integer
      patient_name:
        description: ФИО пациента
        example: Иванов Иван
        type: string
      status:
        description: Текстовый статус приема
        example: Запланирован
        type: string
    type: object
...
  models.UpdatePatientRequest:
    description: Данные для обновления информации о пациенте
    properties:
      birth_date:
        description: Дата рождения
        example: "1980-05-15"
        type: string
      full_name:
        description: ФИО пациента
        example: Смирнов Алексей Петрович
        type: string
      id:
        description: ID пациента
        example: 10
        type: integer
      is_male:
        description: Пол (true - мужской)
        example: true
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: API для работы с приёмами пациентов
  title: ClinicHub API
  version: 1.0.0
paths:
...
  /patient:
    post:
      consumes:
      - application/json
      description: Создает нового пациента с персональными и контактными данными
      parameters:
      - description: Данные пациента
        in: body
        name: info
        required: true
        schema:
          $ref: '#/definitions/models.CreatePatientRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Созданный пациент
          schema:
            $ref: '#/definitions/entities.Patient'
        "400":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/handlers.ResultError'
        "422":
          description: Ошибка валидации данных
          schema:
            $ref: '#/definitions/handlers.ResultError'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handlers.ResultError'
      summary: Создать нового пациента
      tags:
      - Patient
    put:
      consumes:
      - application/json
      description: Обновляет информацию о пациенте
      parameters:
      - description: Данные для обновления
        in: body
        name: info
        required: true
        schema:
          $ref: '#/definitions/models.UpdatePatientRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Обновленный пациент
          schema:
            $ref: '#/definitions/entities.Patient'
        "400":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/handlers.ResultError'
        "404":
          description: Пациент не найден
          schema:
            $ref: '#/definitions/handlers.ResultError'
        "422":
          description: Ошибка валидации
          schema:
            $ref: '#/definitions/handlers.ResultError'
        "500":
          description: Внутренняя ошибка
          schema:
            $ref: '#/definitions/handlers.ResultError'
      summary: Обновить данные пациента
      tags:
      - Patient
  /patient/{id}:
    delete:
      consumes:
      - application/json
      description: Удаляет пациента по ID
      parameters:
      - description: ID пациента
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Успешное удаление
          schema:
            $ref: '#/definitions/handlers.ResultResponse'
        "400":
          description: Некорректный ID
          schema:
            $ref: '#/definitions/handlers.ResultError'
        "404":
          description: Пациент не найден
          schema:
            $ref: '#/definitions/handlers.ResultError'
        "500":
          description: Внутренняя ошибка
          schema:
            $ref: '#/definitions/handlers.ResultError'
      summary: Удалить пациента
      tags:
      - Patient
    get:
      consumes:
      - application/json
      description: Возвращает полную информацию о пациенте
      parameters:
      - description: ID пациента
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Информация о пациенте
          schema:
            $ref: '#/definitions/entities.Patient'
        "400":
          description: Некорректный ID
          schema:
            $ref: '#/definitions/handlers.ResultError'
        "404":
          description: Пациент не найден
          schema:
            $ref: '#/definitions/handlers.ResultError'
        "500":
          description: Внутренняя ошибка
          schema:
            $ref: '#/definitions/handlers.ResultError'
      summary: Получить пациента по ID
      tags:
      - Patient
...
  /reception/{id}:
    get:
      consumes:
      - application/json
      description: Возвращает информацию о приёме
      parameters:
      - description: ID приёма
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Информация о приёме
          schema:
            $ref: '#/definitions/entities.ReceptionHospital'
        "400":
          description: Некорректный ID
          schema:
            $ref: '#/definitions/handlers.ResultError'
        "404":
          description: Приём не найден
          schema:
            $ref: '#/definitions/handlers.ResultError'
        "500":
          description: Внутренняя ошибка
          schema:
            $ref: '#/definitions/handlers.ResultError'
      summary: Получить приём по ID
      tags:
      - Reception
  /receptions/doctor/{doctor_id}:
    get:
      consumes:
      - application/json
      description: Get paginated list of receptions for specific doctor and date
      parameters:
      - description: Doctor ID
        in: path
        name: doctor_id
        required: true
        type: integer
      - description: Date in YYYY-MM-DD format
        in: query
        name: date
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ReceptionShortResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get receptions by doctor and date
      tags:
      - receptions
swagger: "2.0"